% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adopt_Make_Distribution_Plot.R
\name{adopt_Make_Distribution_Plot}
\alias{adopt_Make_Distribution_Plot}
\title{Create a rose plot for two compounds and distributions}
\usage{
adopt_Make_Distribution_Plot(
  compound_names = c("diquat", "glyphosate"),
  data = adopt_hpli
)
}
\arguments{
\item{compound_names}{Vector of desired compounds, length one or more.}

\item{data}{The adopt_hpli dataset.}
}
\value{
A paired rose plot and distribution
}
\description{
Create a rose plot for two compounds and distributions
}
