% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adopt_Make_Paired_Ridge_Plots.R
\name{adopt_Make_Paired_Ridge_Plots}
\alias{adopt_Make_Paired_Ridge_Plots}
\title{Create a ridge plot of the six metrics for a given strategy}
\usage{
adopt_Make_Paired_Ridge_Plots(data = adopt_example, betas = adopt_betas)
}
\arguments{
\item{data}{The questionaire data template, filled in.}

\item{betas}{A dataset defining the beta distributions for confidence levels.}
}
\value{
A plot
}
\description{
Create a ridge plot of the six metrics for a given strategy
}
