% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adopt_Make_Paired_Rose_Plots.R
\name{adopt_Make_Paired_Rose_Plots}
\alias{adopt_Make_Paired_Rose_Plots}
\title{Create a rose plot of the four categories for a given compound}
\usage{
adopt_Make_Paired_Rose_Plots(
  compound_name1 = "diquat",
  compound_name2 = "glyphosate",
  data = adopt_hpli
)
}
\arguments{
\item{compound_name1}{Name of desired compound.}

\item{compound_name2}{Name of desired compound to compare.}

\item{data}{The adopt_hpli dataset.}
}
\value{
A paired rose plot
}
\description{
Create a rose plot of the four categories for a given compound
}
