% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adopt_Make_Rose_and_Dist_Plot.R
\name{adopt_Make_Rose_and_Dist_Plot}
\alias{adopt_Make_Rose_and_Dist_Plot}
\title{Create a rose plot of the four categories for a given compound}
\usage{
adopt_Make_Rose_and_Dist_Plot(compound_name = "diquat", data = adopt_hpli)
}
\arguments{
\item{compound_name}{Name of desired compound.}

\item{data}{The adopt_hpli dataset.}
}
\value{
A rose plot
}
\description{
Create a rose plot of the four categories for a given compound
}
