% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_overall_utility.R
\name{calc_overall_utility}
\alias{calc_overall_utility}
\title{Calcuate the overall utility of the strategies with user-defined weighting}
\usage{
calc_overall_utility(data = opat_example, betas = opat_betas, nsim = 100)
}
\arguments{
\item{data}{The questionaire data template, filled in.}

\item{betas}{A dataset defining the beta distributions for confidence levels.}
}
\value{
A dataframe
}
\description{
Calcuate the overall utility of the strategies with user-defined weighting
}
