% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dot_plot.R
\name{make_dot_plot}
\alias{make_dot_plot}
\title{Create a dot plot summarizing change in the six metrics}
\usage{
make_dot_plot(data = opat_example, betas = opat_betas)
}
\arguments{
\item{data}{The questionaire data template, filled in.}

\item{betas}{A dataset defining the beta distributions for confidence levels.}
}
\value{
A plot
}
\description{
Create a dot plot summarizing change in the six metrics
}
