% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ridge_plots.R
\name{make_ridge_plots}
\alias{make_ridge_plots}
\title{Create a ridge plot of the six metrics for a given strategy}
\usage{
make_ridge_plots(data = opat_example, betas = opat_betas)
}
\arguments{
\item{data}{The questionaire data template, filled in.}

\item{betas}{A dataset defining the beta distributions for confidence levels.}
}
\value{
A plot
}
\description{
Create a ridge plot of the six metrics for a given strategy
}
