% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ridge_plots_pair.R
\name{make_ridge_plots_pair}
\alias{make_ridge_plots_pair}
\title{Create overlapping ridge plots of the six metrics, comparing each strategy}
\usage{
make_ridge_plots_pair(data = opat_example, betas = opat_betas)
}
\arguments{
\item{data}{The questionaire data template, filled in.}

\item{betas}{A dataset defining the beta distributions for confidence levels.}
}
\value{
A plot
}
\description{
Create overlapping ridge plots of the six metrics, comparing each strategy
}
