% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rose_plot.R
\name{make_rose_plot}
\alias{make_rose_plot}
\title{Create a rose plot of the four categories for a given compound}
\usage{
make_rose_plot(compound_name, data)
}
\arguments{
\item{compound_name}{Name of desired compound.}

\item{data}{The opat_hpli dataset.}
}
\value{
A rose plot
}
\description{
Create a rose plot of the four categories for a given compound
}
