% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_rose_plot_pair.R
\name{make_rose_plot_pair}
\alias{make_rose_plot_pair}
\title{Create a rose plot of the four categories for a given compound}
\usage{
make_rose_plot_pair(
  compound_name1 = "diquat",
  compound_name2 = "glyphosate",
  data = opat_hpli
)
}
\arguments{
\item{compound_name1}{Name of desired compound.}

\item{compound_name2}{Name of desired compound to compare.}

\item{data}{The opat_hpli dataset.}
}
\value{
A paired rose plot
}
\description{
Create a rose plot of the four categories for a given compound
}
