% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adopt_Make_Paired_Ridge_Plots.R
\name{adopt_Make_Paired_Ridge_Plots}
\alias{adopt_Make_Paired_Ridge_Plots}
\title{Create overlapping ridge plots of the six metrics, comparing each strategy}
\usage{
adopt_Make_Paired_Ridge_Plots(data = adopt_example, betas = opat_betas)
}
\arguments{
\item{data}{The questionaire data template, filled in.}

\item{betas}{A dataset defining the beta distributions for confidence levels.}
}
\value{
A plot
}
\description{
Create overlapping ridge plots of the six metrics, comparing each strategy
}
